usb: pd_engine: free struct psy_changed_event
In psy_changed(), the kernel memory is allocated for struct
psy_changed_event, which represents the work to be queued in
the driver's work queue. However, in the handler psy_changed_handler(),
this struct is not freed. This causes a memory leakage.

Freeing the struct at the end of the handler to fix the issue.

Bug: 33429649
Bug: 67648334
Change-Id: I84ef6b4055cbfc2d03679376cae780a1ce0576e6
Signed-off-by: Badhri Jagan Sridharan <Badhri@google.com>
Signed-off-by: Eliminater74 <eliminater74@gmail.com>