ANDROID: tracing/ring_buffer: Try harder to allocate
Ftrace fails to allocate large buffers on walleye/taimen due to large
number of CPUs coupled with effects of large amounts of caching in the
page cache.  Currently the ring buffer allocation doesn't retry after
direct-reclaim made some progress but still wasn't able to find a free
page. On retrying I see that the allocations almost always succeed.  The
retry doesn't happen because GFP_NORETRY is used in the tracer to
prevent the case where we might OOM, however this is incorrect in the
sense a RETRY doesn't necessarily mean it will OOM (as seen in my
testing, most of the time the retry just needs to loop back and find a
free page). In order to retry but still prevent the OOM, I re-use the
PF_DUMPCORE flag which bails out incase we are about to OOM.

Test:
* systrace -b 64000 and -b 100000 is working after dropping GFP_NORETRY.
* Used memeater to use 3GB memory and -b 64000 and saw that OOM didn't
trigger with PF_DUMPCORE (and triggered without PF_DUMPCORE).

Bug: 62918608
Change-Id: I85de23e843d2a4fdf2276905b1ae8af4486b3bd7
Signed-off-by: Joel Fernandes <joelaf@google.com>
Signed-off-by: Eliminater74 <eliminater74@gmail.com>